requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tencentcloud-sdk-python==3.0.1035
pillow==10.2.0
selenium==4.18.1
webdriver-manager==4.0.1
openai>=1.12.0
python-ffmpeg==2.0.4
pytest==8.0.0
pytest-asyncio==0.23.5
websocket-client==1.7.0
pyJianYingDraft>=1.0.0 